3409 Require scope
